% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banovaplot.R
\name{banovaplot}
\alias{banovaplot}
\title{Bayasian ANOVA plot}
\usage{
banovaplot(
  baov,
  ylim = NULL,
  xlim = NULL,
  horizontal = TRUE,
  cols = NULL,
  labels = "default",
  xlab = "",
  ylab = "",
  pch = 16,
  lwd = 2,
  cex = 1.5,
  las = NA,
  cex.axis = 1,
  grid = TRUE,
  robust = FALSE,
  yaxs = "r",
  xaxs = "r",
  line = TRUE,
  ...
)
}
\arguments{
\item{baov}{--.}

\item{ylim}{--.}

\item{xlim}{--.}

\item{horizontal}{--.}

\item{cols}{--.}

\item{labels}{--.}

\item{xlab}{--.}

\item{ylab}{--.}

\item{pch}{--.}

\item{lwd}{--.}

\item{cex}{--.}

\item{las}{--.}

\item{cex.axis}{--.}

\item{grid}{--.}

\item{robust}{--.}

\item{yaxs}{--.}

\item{xaxs}{--.}

\item{line}{--.}

\item{...}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
