% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmplot.R
\name{brmplot}
\alias{brmplot}
\title{Plot brm parameter information}
\usage{
brmplot(
  mat,
  ylim = NULL,
  xlim = NULL,
  horizontal = TRUE,
  add = FALSE,
  xs = NULL,
  col = 1,
  labels = "default",
  xlab = "",
  ylab = "",
  pch = 16,
  lwd = 2,
  cex = 1.5,
  las = NA,
  cex.axis = 1,
  grid = TRUE,
  cex.lab = 1,
  yaxs = "r",
  xaxs = "r",
  line = TRUE,
  nudge = 0,
  omit = 0,
  ...
)
}
\arguments{
\item{mat}{a dataframe or matrix representing the posterior summary of brms model parameters.}

\item{ylim}{optionally override the plot y range.}

\item{xlim}{optionally override the plot x range.}

\item{horizontal}{if TRUE, plot parameters across the x axis.}

\item{add}{if TRUE, add to existing plot.}

\item{xs}{provide x axis location for plotting parameters.}

\item{col}{color for points and intervals.}

\item{labels}{a vector of labels. If "", labels are blank.}

\item{xlab}{optionally override the plot x axis range.}

\item{ylab}{optionally override the plot y axis range.}

\item{pch}{points for parameter means.}

\item{lwd}{line widths for intervals.}

\item{cex}{point size for parameter means.}

\item{las}{orientation of axis labels.}

\item{cex.axis}{expansion of axis tick numbers.}

\item{grid}{if TRUE, a grid is added.}

\item{cex.lab}{expansion of axis labels.}

\item{yaxs}{y axis edge settings.}

\item{xaxs}{x axis edge settings.}

\item{line}{if TRUE, add a line at x=0.}

\item{nudge}{nudge points along x axis by this amount.}

\item{omit}{a vector of parameters to be omitted.}

\item{...}{additional parameters are passed to internal call of brmplot.}
}
\description{
Make plots of brm model parameter means and intervals.
}
\examples{
\dontrun{
model_height_vtl = brms::get_model("11_model_height_vtl_f0")
fixed_effects = fixef (model_height_vtl)
brmplot (fixed_effects, omit = 1)
}

}
