% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cols}
\alias{cols}
\title{Colors for plotting}
\format{
An object of class \code{character} of length 10.
}
\usage{
cols
}
\description{
A vector of RGB values for colors used for making the plots in the book: light pink,lavender, coral, yellow, deep green, teal, sky blue, maroon, deep purple, dark orange.
}
\keyword{datasets}
