% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_factors.R
\name{divide_factors}
\alias{divide_factors}
\title{Divide factors}
\usage{
divide_factors(model, formula = NULL)
}
\arguments{
\item{model}{--.}

\item{formula}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
