% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model.R
\name{download_model}
\alias{download_model}
\title{Download pre-fit models}
\usage{
download_model(model_name)
}
\arguments{
\item{model_name}{the file name to download. Each file is named like this "chapterNumber_modelName".}
}
\value{
the pre-fit brm model, loaded from an RDS file downloaded form GitHub.
}
\description{
Models are downloaded from the book GitHub page.
}
\examples{
\dontrun{
download_model ("2_model.RDS")
}

}
