% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_data_all}
\alias{exp_data_all}
\title{Data from a listening experiment.}
\format{
An object of class \code{data.frame} with 4170 rows and 13 columns.
}
\usage{
exp_data_all
}
\description{
The data from the listening experiment analyzed in "Bayesian multilevel models for repeated-measures data".
}
\section{Columns}{

\describe{
\item{L}{An integer from 1-15 indicating which listener responded to the trial.}
\item{C}{A letter representing the apparent speaker category (b=boy, g=girl, m=man, w=woman) reported by the listener for each trial.}
\item{height}{A floating-point number representing the height (in centimeters) reported for the speaker on each trial.}
\item{R}{A letter representing the resonance scaling for the stimulus on each trial. The coding is a (actual) for the unmodified resonance and b (big) for the modified resonance (intended to sound bigger)}
\item{S}{An integer from 1-139 indicating which speaker produced the trial stimulus.}
\item{C_v}{A letter representing the veridical (actual) speaker category (b=boy, g=girl, m=man, w=woman) for each trial.}
\item{vtl}{An estimate of the speaker’s vocal-tract length in centimeters.}
\item{f0}{The vowel fundamental frequency (f0) measured in Hertz.}
\item{dur}{The duration of the vowel sound, in milliseconds.}
\item{G}{The apparent gender of the speaker indicated by the listener, f (female) or m (male).}
\item{A}{The apparent age of the speaker indicated by the listener, a (adult) or c (child).}
\item{G_v}{The veridical gender of the speaker indicated by the listener, f (female) or m (male).}
\item{A_v}{The veridical age of the speaker indicated by the listener, a (adult) or c (child).}
}
}

\keyword{datasets}
