% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersection.R
\name{find_intersection}
\alias{find_intersection}
\title{Find intersection of lines or planes
--}
\usage{
find_intersection(coeffs1, coeffs2, z = 0)
}
\arguments{
\item{coeffs1}{--.}

\item{coeffs2}{--.}

\item{z}{--.}
}
\description{
Find intersection of lines or planes
--
}
\examples{
\dontrun{
 # coming soon
}


}
