% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findauc.R
\name{findauc}
\alias{findauc}
\title{Empirically estimate the area under the curve}
\usage{
findauc(x, density, log = FALSE)
}
\arguments{
\item{x}{--.}

\item{density}{--.}

\item{log}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
