% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forpaper.R
\name{forpaper}
\alias{forpaper}
\title{Make output pastable into paper}
\usage{
forpaper(x, digits = 2)
}
\arguments{
\item{x}{--.}

\item{digits}{--.}
}
\description{
--
}
\examples{
\dontrun{
 # coming soon
}
}
