% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priors.R
\name{get_prior}
\alias{get_prior}
\title{Get priors in brms models
--}
\usage{
get_prior(formula, data, family = gaussian(), ...)
}
\arguments{
\item{formula}{--.'}

\item{data}{--.}

\item{family}{--.}

\item{...}{--.}
}
\description{
Get priors in brms models
--
}
\examples{
\dontrun{
 # coming soon
}


}
