% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_check.R
\name{get_samples}
\alias{get_samples}
\title{Plot posterior samples in a data frame#'
--}
\usage{
get_samples(model)
}
\arguments{
\item{model}{--.}
}
\description{
Plot posterior samples in a data frame#'
--
}
\examples{
\dontrun{
 # coming soon
}

}
