% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcorrs.R
\name{getcorrs}
\alias{getcorrs}
\alias{get_corrs}
\title{Get random effect correlations}
\usage{
getcorrs(model, factor)

get_corrs(model, factor)
}
\arguments{
\item{model}{--.}

\item{factor}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
