% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getheadings.R
\name{getheadings}
\alias{getheadings}
\title{Get rmd file headings
--}
\usage{
getheadings(file)
}
\arguments{
\item{file}{--.}
}
\description{
Get rmd file headings
--
}
\examples{
\dontrun{
 # coming soon
}

}
