% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{getsamples}
\alias{getsamples}
\alias{get_samples}
\title{Get posterior samples}
\usage{
getsamples(model)

get_samples(model)
}
\arguments{
\item{model}{The brms model.}
}
\description{
Gets posterior samples from a brms model, returns these as a dataframe.
}
\examples{
\dontrun{
model_height_vtl = bmmb::get_model("11_model_height_vtl_f0")
samples = bmmb::get_samples (model_height_vtl)
samples[1:6,1:10]
}

}
