% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sds.R
\name{getsds}
\alias{getsds}
\alias{get_sds}
\title{Get random effect standard deviations}
\usage{
getsds(model)

get_sds(model)
}
\arguments{
\item{model}{The brms model.}
}
\description{
Gets summaries of the random effect standard deviations from a brms model.
}
\examples{
\dontrun{
model_height_vtl = bmmb::get_model("11_model_height_vtl_f0")

# Standard deviations of different groups of random effects
bmmb::get_sds (model_height_vtl)
}

}
