% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{height_data}
\alias{height_data}
\title{Height data}
\format{
An object of class \code{data.frame} with 38 rows and 6 columns.
}
\usage{
height_data
}
\description{
Data regarding the heights of males and females of different ages. From: Fryar, C. D., Gu, Q., & Ogden, C. L. (2012). Anthropometric reference data for children and adults: United States, 2007-2010. Vital and Health Statistics. Series 11, Data from the National Health Survey, 252, 1–48.
}
\section{Columns}{

\describe{
\item{age}{The agre group, in years.}
\item{gender}{The group gender, male or female}
\item{height}{Height, in centimeters}
\item{sesm}{The standard error of the sample mean.}
\item{sd}{The standard deviation, calculated using the sesm and sample size}
\item{sample}{The size of the sample for each agre and gender group.}
}
}

\keyword{datasets}
