% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.interaction.R
\name{interactionplot}
\alias{interactionplot}
\title{Plot interactions}
\usage{
interactionplot(
  x.factor,
  trace.factor,
  response,
  fun = mean,
  type = c("l", "p", "b", "o", "c"),
  legend = TRUE,
  trace.label = deparse(substitute(trace.factor)),
  fixed = FALSE,
  xlab = deparse(substitute(x.factor)),
  ylab = ylabel,
  ylim = range(cells, na.rm = TRUE),
  lty = nc:1,
  col = 1,
  pch = c(1L:9, 0, letters),
  xpd = NULL,
  leg.bg = graphics::par("bg"),
  leg.bty = "n",
  xtick = FALSE,
  xaxt = graphics::par("xaxt"),
  axes = TRUE,
  leg.x = NULL,
  leg.y = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{x.factor}{--.}

\item{trace.factor}{--.}

\item{response}{--.}

\item{fun}{--.}

\item{type}{--.}

\item{legend}{--.}

\item{trace.label}{--.}

\item{fixed}{--.}

\item{xlab}{--.}

\item{ylab}{--.}

\item{ylim}{--.}

\item{lty}{--.}

\item{col}{--.}

\item{pch}{--.}

\item{xpd}{--.}

\item{leg.bg}{--.}

\item{leg.bty}{--.}

\item{xtick}{--.}

\item{xaxt}{--.}

\item{axes}{--.}

\item{leg.x}{--.}

\item{leg.y}{--.}

\item{xlim}{--.}

\item{...}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
