% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logits.R
\name{inverse_logit}
\alias{inverse_logit}
\alias{ztop}
\title{Convert logits to probabilities}
\usage{
inverse_logit(z)

ztop(z)
}
\arguments{
\item{z}{a scalar logit value, or a vector of scalars.}
}
\description{
Convert logits to probabilities
}
\examples{
\dontrun{
--
}


}
