% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee_etal_data}
\alias{lee_etal_data}
\title{Lee et al. data}
\format{
An object of class \code{data.frame} with 300 rows and 11 columns.
}
\usage{
lee_etal_data
}
\description{
Data regarding the average acoustic characteristics of speech of children of different ages. From: Lee, S., Potamianos, A., & Narayanan, S. (1999). Acoustics of children’s speech: Developmental changes of temporal and spectral parameters. The Journal of the Acoustical Society of America, 105(3), 1455–1468.
}
\section{Columns}{

\describe{
\item{vowel}{The vowel category}
\item{f0}{Average f0 for that vowel, for that age and gender group.}
\item{f0sd}{Standard deviation of f0 for that vowel, for that age and gender group.}
\item{f1}{Average f1 for that vowel, for that age and gender group.}
\item{f1sd}{Standard deviation of f1 for that vowel, for that age and gender group.}
\item{f2}{Average f2 for that vowel, for that age and gender group.}
\item{f2sd}{Standard deviation of f2 for that vowel, for that age and gender group.}
\item{f3}{Average f3 for that vowel, for that age and gender group.}
\item{f3sd}{Standard deviation of f3 for that vowel, for that age and gender group.}
\item{age}{The age of the speaker group, in year.}
\item{gender}{The gender of the speaker group, male or female}
}
}

\keyword{datasets}
