% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lightpink}
\alias{lightpink}
\title{Colors for book plots}
\format{
An object of class \code{character} of length 1.
}
\usage{
lightpink
}
\description{
Colors used in book figures.
}
\keyword{datasets}
