% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{mae}
\alias{mae}
\title{Find the median/mean absolute error}
\usage{
mae(x, mean = FALSE)
}
\arguments{
\item{x}{a vector of numbers.}

\item{mean}{if TRUE, the mean absolute deviation is returned.}
}
\description{
The median absolute deviation is found by default.
}
