% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Make territorial map from planes
--}
\usage{
make_map(
  parameters,
  xlim = NULL,
  ylim = NULL,
  density = 1000,
  plot = FALSE,
  new = TRUE,
  points = NULL,
  ...
)
}
\arguments{
\item{parameters}{--.}

\item{xlim}{--.}

\item{ylim}{--.}

\item{density}{--.}

\item{plot}{--.}

\item{new}{--.}

\item{points}{--.}

\item{...}{--.}
}
\description{
Make territorial map from planes
--
}
\examples{
\dontrun{
 # coming soon
}


}
