% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebook.R
\name{makebook}
\alias{makebook}
\title{Make word book}
\usage{
makebook(folder = "./", output_format = "bookdown::word_document2")
}
\arguments{
\item{folder}{--.}

\item{output_format}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
