% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedoc.R
\name{makedoc}
\alias{makedoc}
\title{Make word document}
\usage{
makedoc(file, output_format = "bookdown::word_document2")
}
\arguments{
\item{file}{--.}

\item{output_format}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
