% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_check.R
\name{p_check}
\alias{p_check}
\title{Plot predictive checks}
\usage{
p_check(
  model,
  show_data = FALSE,
  n_samples = 10,
  samples = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  ignore_extreme = TRUE,
  re_formula = NULL,
  ...
)
}
\arguments{
\item{model}{--.}

\item{show_data}{--.}

\item{n_samples}{--.}

\item{samples}{--.}

\item{xlim}{--.}

\item{ylim}{--.}

\item{xlab}{--.}

\item{ylab}{--.}

\item{ignore_extreme}{--.}

\item{re_formula}{--.}

\item{...}{--.}
}
\description{
--
}
\examples{
\dontrun{
 # coming soon
}
}
