% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot territorial maps}
\usage{
plot_map(maps, new = TRUE, colors = NULL, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{maps}{the output of bmmb::make_map.}

\item{new}{if TRUE, a new plot is created.}

\item{colors}{a vector of colors, one for each territory.}

\item{xlim}{plot x axis limits.}

\item{ylim}{plot y axis limits.}

\item{...}{additional parameters are passed to the internal call of plot().}
}
\description{
Plot territorial maps estimated using the bmmb::make_map function provided in this package. See chapter 12 for usage examples.
}
