% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R, R/plotcols.R
\name{plotcols}
\alias{plotcols}
\title{Plot plotting colors}
\usage{
plotcols(cols = NA)

plotcols(cols = NA)
}
\arguments{
\item{cols}{--.}
}
\description{
--

--
}
\examples{
\dontrun{
--
}


\dontrun{
--
}


}
