% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_check.R
\name{predict_n}
\alias{predict_n}
\title{Predictive one version of dependent variables}
\usage{
predict_n(model, n = 1, draw_ids = NULL, ...)
}
\arguments{
\item{model}{brms model.}

\item{n}{the number of samples to get.}

\item{draw_ids}{a vector of specific samples if desired.}

\item{...}{additional parameters passed to internal call of predict().}
}
\description{
Get posterior predictions.
}
\examples{
\dontrun{
 # coming soon
}
}
