% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_summary.R
\name{prior_summary}
\alias{prior_summary}
\title{Summarize priors in brms models
--}
\usage{
prior_summary(model, all = TRUE)
}
\arguments{
\item{model}{--.#'}

\item{all}{--.}
}
\description{
Summarize priors in brms models
--
}
\examples{
\dontrun{
 # coming soon
}


}
