% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tns.R
\name{qtns}
\alias{qtns}
\title{Quantile function for non-standardized t distribution}
\usage{
qtns(x, m = 0, s = 1, df)
}
\arguments{
\item{x}{vector of quantiles.}

\item{m}{distribution mean.}

\item{s}{distribution scale.}

\item{df}{degrees of freedom.}
}
\description{
Quantile function for non-standardized t distribution
}
