% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_bayes.R
\name{r2_bayes}
\alias{r2_bayes}
\alias{r2bayes}
\title{Bayesian R2}
\usage{
r2_bayes(model, re_formula = NULL, summary = TRUE)

r2bayes(model, re_formula = NULL, summary = TRUE)
}
\arguments{
\item{model}{brms hypothesis object.}

\item{re_formula}{--.}

\item{summary}{--.}
}
\description{
R2 with and without random effects.
}
\examples{
\dontrun{
 # coming soon
}
}
