% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tns.R
\name{rbernoulli}
\alias{rbernoulli}
\title{Random generation of Bernoulli variable}
\usage{
rbernoulli(n, p)
}
\arguments{
\item{n}{number of observations.}

\item{p}{probability of a success.}
}
\description{
Random generation of Bernoulli variable
}
