% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_hypothesis.R
\name{short_hypothesis}
\alias{short_hypothesis}
\title{Short hypothesis printout}
\usage{
short_hypothesis(hypothesis)
}
\arguments{
\item{hypothesis}{brms hypothesis object.}
}
\description{
Shorter brms hypothesis print statement.
}
\examples{
\dontrun{
 # coming soon
}
}
