% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatemodel-LS-200003-LIN.R, R/updatemodel.R
\name{updatemodel}
\alias{updatemodel}
\title{update model}
\usage{
updatemodel(model, chapter = 0)

updatemodel(model, chapter = 0)
}
\arguments{
\item{predictions}{--.}

\item{n_samples}{--.}
}
\description{
--

--
}
\examples{
\dontrun{
 # coming soon
}
\dontrun{
 # coming soon
}
}
