% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcount.R
\name{wordcounts}
\alias{wordcounts}
\title{Count words in Rmd file excluding code chunks.}
\usage{
wordcounts(folder = getwd())
}
\arguments{
\item{folder}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
