% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banova.R
\name{banova}
\alias{banova}
\title{Bayesian ANOVA}
\usage{
banova(model, superpopulation = FALSE, collapse = TRUE)
}
\arguments{
\item{model}{--.}

\item{superpopulation}{--.}

\item{collapse}{--.}
}
\description{
--
}
\examples{
\dontrun{
--
}


}
