% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banovaplot.R
\name{banovaplot}
\alias{banovaplot}
\title{Bayasian ANOVA plot}
\usage{
banovaplot(
  baov,
  ylim = NULL,
  xlim = NULL,
  horizontal = TRUE,
  cols = NULL,
  labels = "default",
  xlab = "",
  ylab = "",
  pch = 16,
  lwd = 2,
  cex = 1.5,
  las = NA,
  cex.axis = 1,
  grid = TRUE,
  robust = FALSE,
  yaxs = "r",
  xaxs = "r",
  line = TRUE,
  ...
)
}
\arguments{
\item{baov}{The dataframe representing a Bayesian Analysis of Variance.}

\item{ylim}{optionally override the plot y range.}

\item{xlim}{optionally override the plot x range.}

\item{horizontal}{if TRUE, plot parameters across the x axis.}

\item{cols}{A vector of colors used to plot the different clusters.}

\item{labels}{a vector of labels. If "", labels are blank.}

\item{xlab}{optionally override the plot x axis range.}

\item{ylab}{optionally override the plot y axis range.}

\item{pch}{points for parameter means.}

\item{lwd}{line widths for intervals.}

\item{cex}{point size for parameter means.}

\item{las}{orientation of axis labels.}

\item{cex.axis}{expansion of axis tick numbers.}

\item{grid}{If TRUE a grid is added to the plot.}

\item{robust}{if TRUE, robust summaries of posterior distributions are used.}

\item{yaxs}{y axis edge settings.}

\item{xaxs}{x axis edge settings.}

\item{line}{if TRUE, add a line at x=0.}

\item{...}{additional parameters are passed to internal call of bmmb::brmplot().}
}
\description{
A function to plot the output of the banova function included in this package.
}
\examples{
\dontrun{
model_height_vtl = bmmb::get_model("11_model_height_vtl_f0")
options (contrasts = c('contr.sum','contr.sum'))
baov = bmmb::banova (model_height_vtl)
bmmb::banovaplot (baov)
}


}
