% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmplot.R
\name{brmplot}
\alias{brmplot}
\title{Plot brm parameters and intervals}
\usage{
brmplot(
  mat,
  ylim = NULL,
  xlim = NULL,
  horizontal = TRUE,
  add = FALSE,
  xs = NULL,
  col = 1,
  labels = "default",
  xlab = "",
  ylab = "",
  pch = 16,
  lwd = 2,
  cex = 1.5,
  las = NA,
  cex.axis = 1,
  grid = TRUE,
  cex.lab = 1,
  robust = FALSE,
  yaxs = "r",
  xaxs = "r",
  line = TRUE,
  nudge = 0,
  omit = 0,
  ...
)
}
\arguments{
\item{mat}{a dataframe or matrix representing the posterior summary of brms model parameters.}

\item{ylim}{optionally override the plot y range.}

\item{xlim}{optionally override the plot x range.}

\item{horizontal}{--.}

\item{add}{--.}

\item{xs}{--.}

\item{col}{--.}

\item{labels}{--.}

\item{xlab}{--.}

\item{ylab}{--.}

\item{pch}{--.}

\item{lwd}{--.}

\item{cex}{--.}

\item{las}{--.}

\item{cex.axis}{--.}

\item{grid}{--.}

\item{cex.lab}{--.}

\item{robust}{--.}

\item{yaxs}{--.}

\item{xaxs}{--.}

\item{line}{--.}

\item{nudge}{--.}

\item{omit}{--.}

\item{...}{--.}
}
\value{
--.
}
\description{
Make plots of brm model parameter means and intervals.
}
\examples{
\dontrun{
 # coming soon
}

}
