% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_paper.R
\name{forpaper}
\alias{forpaper}
\alias{for_paper}
\title{Format table output for copying into a paper}
\usage{
forpaper(x, digits = 2)

for_paper(x, digits = 2)
}
\arguments{
\item{x}{--.}

\item{digits}{--.}
}
\description{
Format table output for copying into a paper
}
\examples{
\dontrun{
model_height_vtl = brms::get_model("11_model_height_vtl_f0")
fixed_effects = fixef (model_height_vtl)
forpaper (fixed_effects)
}
}
