% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Download pre-fit models}
\usage{
get_model(model_name)
}
\arguments{
\item{model_name}{the file name to download. Models are named like this "chapterNumber_modelName.RDS", based on their name in the book.}
}
\value{
Returns the pre-fit brms model, loaded from an RDS file downloaded form GitHub.
}
\description{
Download models for the book "Bayesian Multilevel Models for Repeated Measures Data" from the book GitHub page: https://github.com/santiagobarreda/bmmrmd/.
}
\examples{
\dontrun{
# Get the first model from chapter 3 like this:
get_model ("3_model.RDS")

# And the multinomial model from chapter 12 like this:
get_model ("12_model_multinomial.RDS")
}

}
