% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priors.R
\name{get_prior}
\alias{get_prior}
\title{Get priors in brms models}
\usage{
get_prior(formula, data, family = "gaussian", ...)
}
\arguments{
\item{formula}{A description of the model to be fit.'}

\item{data}{The data to be used by the model.}

\item{family}{Information about the response distribution and/or link function.}

\item{...}{additional arguments are passed to brms::get_prior.}
}
\description{
Information about priors which may be set for a brms model. This just calls brms::get_prior but changes the output of the formatting slightly to allow for columns to be imitted from the output. Please see the help file for brms::get_prior for more information.
}
\examples{
\dontrun{
brms::get_prior (brms::bf(height ~ A*G + (A*G|L) + (1|S),
                          sigma ~ A),
                 data = exp_data, family="student")
}

}
