% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsamples.R
\name{getsamples}
\alias{getsamples}
\alias{get_samples}
\title{Plot posterior samples in a data frame#'
--}
\usage{
getsamples(model)

get_samples(model)
}
\arguments{
\item{model}{--.}
}
\description{
Plot posterior samples in a data frame#'
--
}
\examples{
\dontrun{
 # coming soon
}

}
