% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee_etal_data}
\alias{lee_etal_data}
\title{Lee et al. data}
\format{
An object of class \code{data.frame} with 300 rows and 11 columns.
}
\source{
\url{https://github.com/santiagobarreda/FastTrackR}
}
\usage{
lee_etal_data
}
\description{
The Hillenbrand et al. (1995) data. The formant columns are so that column FXY represents the frequency for the Xth formant, for the Yth time slice.
}
\section{Columns}{

\describe{
\item{file}{the recording file name.}
\item{f0}{average f0 in Hertz.}
\item{duration}{vowel duration in milliseconds.}
\item{label}{The label used for plotting the vowel.}
\item{group}{A group number, used for plotting}
\item{color}{The color Praat will use for plotting.}
\item{number}{The file number.}
\item{f11}{The average frequency for the first formant, for the first time slice.}
\item{f21}{The average frequency for the second formant, for the first time slice.}
\item{...}{And so on.}
}
}

\references{
Hillenbrand, J.M., Getty, L.A., Clark, M.J., and Wheeler, K. (1995). "Acoustic characteristics of American English vowels," Journal of the Acoustical Society of America, 97, 3099-3111.
}
\keyword{datasets}
