% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkjcorrdensity.R
\name{lkjcorrdensity}
\alias{lkjcorrdensity}
\title{LKJ correlation prior density}
\usage{
lkjcorrdensity(eta = 2)
}
\arguments{
\item{eta}{--.}
}
\description{
--
}
\examples{
\dontrun{
 # coming soon
}
}
