% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Make territorial map from planes}
\usage{
make_map(
  parameters,
  xlim = NULL,
  ylim = NULL,
  density = 1000,
  plot = FALSE,
  new = TRUE,
  points = NULL,
  ...
)
}
\arguments{
\item{parameters}{a matrix containing the parameters of the surfaces representing each category.}

\item{xlim}{the x axis range for boundary calculations.}

\item{ylim}{the x axis range for boundary calculations.}

\item{density}{the number of points sampled along the x and y axis ranges.}

\item{plot}{if TRUE, the results are plotted.}

\item{new}{if TRUE, a new plot is created.}

\item{points}{if a set of x and y axis points are provided, plot ranges are based on the area covered by the points.}

\item{...}{additional parameters are passed to internal call of bmmb::plot_map().}
}
\description{
Make territorial maps estimated based on multinomial regression models. See chapter 12 for usage examples and a more detailed description.
}
