% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_check.R
\name{p_check}
\alias{p_check}
\title{Plot predictive checks}
\usage{
p_check(
  predictions,
  n_samples = 10,
  samples = NULL,
  xlim = NULL,
  ylim = NULL,
  ignore_extreme = TRUE,
  ...
)
}
\arguments{
\item{predictions}{--.}

\item{n_samples}{--.}

\item{samples}{--.}

\item{xlim}{--.}

\item{ylim}{--.}

\item{ignore_extreme}{--.}

\item{...}{--.}
}
\description{
--
}
\examples{
\dontrun{
 # coming soon
}
}
