% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot territorial maps
--}
\usage{
plot_map(maps, new = TRUE, colors = NULL, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{maps}{--.}

\item{new}{--.}

\item{colors}{--.}

\item{xlim}{--.}

\item{ylim}{--.}

\item{...}{--.}
}
\description{
Plot territorial maps
--
}
\examples{
\dontrun{
 # coming soon
}


}
