% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_check.R
\name{predict_n}
\alias{predict_n}
\title{Predictive one version of dependent variables}
\usage{
predict_n(model, n = 1, draw_ids = NULL, ...)
}
\arguments{
\item{model}{--.}

\item{draw_ids}{--.}

\item{...}{--.}
}
\description{
--
}
\examples{
\dontrun{
 # coming soon
}
}
