% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_summary.R
\name{prior_summary}
\alias{prior_summary}
\title{Summarize priors in brms models
Information about priors from an existing brms model. This just calls brms::prior_summary but changes the output of the formatting slightly to allow for columns to be imitted from the output. Please see the help file for brms::prior_summary for more information.}
\usage{
prior_summary(model)
}
\arguments{
\item{model}{The brms model.}
}
\description{
Summarize priors in brms models
Information about priors from an existing brms model. This just calls brms::prior_summary but changes the output of the formatting slightly to allow for columns to be imitted from the output. Please see the help file for brms::prior_summary for more information.
}
\examples{
\dontrun{
model_height_vtl = bmmb::get_model("11_model_height_vtl_f0")
bmmb::prior_summary (model_height_vtl)
}


}
