% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tns.R
\name{ptns}
\alias{ptns}
\title{Distribution function for non-standardized t distribution}
\usage{
ptns(x, m = 0, s = 1, df)
}
\arguments{
\item{x}{--.}

\item{m}{--.}

\item{s}{--.}

\item{df}{--.}
}
\description{
Distribution function for non-standardized t distribution
}
\examples{
\dontrun{
 # coming soon
}
}
