% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tns.R
\name{qtns}
\alias{qtns}
\title{Quantile function for non-standardized t distribution}
\usage{
qtns(x, m = 0, s = 1, df)
}
\arguments{
\item{x}{--.}

\item{m}{--.}

\item{s}{--.}

\item{df}{--.}
}
\description{
Quantile function for non-standardized t distribution
}
\examples{
\dontrun{
 # coming soon
}
}
