% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tns.R
\name{rtns}
\alias{rtns}
\title{Random generation of non-standardized t distributed numbers}
\usage{
rtns(n, m = 0, s = 1, df)
}
\arguments{
\item{n}{number of observations}

\item{m}{distribution mean.}

\item{s}{distribution scale.}

\item{df}{degrees of freedom.}
}
\description{
Random generation of non-standardized t distributed numbers
}
