% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_hypothesis.R
\name{short_hypothesis}
\alias{short_hypothesis}
\title{Short hypothesis printout}
\usage{
short_hypothesis(x, ...)
}
\arguments{
\item{x}{the brms model.}

\item{...}{additional arguments passed to brms::hypothesis.}
}
\description{
This is just a clone of brms::hypothesis with a smaller print statement, needed to get output to fit in the width of a book page. Please see the help file for brms::hypothesis for more information.
}
