% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_summary.R
\name{short_summary}
\alias{short_summary}
\title{Short summary printout}
\usage{
short_summary(model)
}
\arguments{
\item{hypothesis}{brms model object.}
}
\description{
Shorter brms model print statement.
}
\examples{
\dontrun{
 # coming soon
}
}
