% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_summary.R
\name{short_summary}
\alias{short_summary}
\title{Short summary printout}
\usage{
short_summary(model, omit_conv_stats = TRUE, omit_corrs = FALSE, digits = 2)
}
\arguments{
\item{model}{--.}

\item{omit_conv_stats}{--.}

\item{omit_corrs}{--.}

\item{digits}{--.}
}
\description{
Shorter brms model print statement.
}
\examples{
\dontrun{
 # coming soon
}
}
