% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatemodel.R
\name{updatemodel}
\alias{updatemodel}
\title{update model}
\usage{
updatemodel(model, chapter = 0)
}
\arguments{
\item{model}{--.}

\item{chapter}{--.}
}
\description{
--
}
\examples{
\dontrun{
 # coming soon
}
}
